Domain layer and getAllPokemon use case created. Code refactored: mediaplayer implementation moved, loading composable moved, pokemon list moved.
