+++ app/src/main/java/com/example/pokedexdroid/data/model/repository/PokemonRepository.kt
@@ -1,0 +1,8 @@
+package com.example.pokedexdroid.data.model.repository
+
+import com.example.pokedexdroid.data.model.PokemonCardData
+
+class PokemonRepository(private val localDataSource: PokemonLocalDataSource) {
+
+    suspend fun getFirstPage(): List<PokemonCardData> = localDataSource.getFirstPokemonPage()
+}
