+++ app/src/main/java/com/example/pokedexdroid/viewmodel/PokemonViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.pokedexdroid.data.RetrofitService
 import com.example.pokedexdroid.data.model.PokemonCardData
 import com.example.pokedexdroid.data.model.repository.PokemonRepository
+import com.example.pokedexdroid.domain.GetAllPokemonUseCase
 import com.example.pokedexdroid.ui.theme.utils.UIState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -14,15 +15,16 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 /**
- * @param repository is unique instance that manages the getting of the pokemon list data.
- * Cause the view model receive a repository as argument is necessary create an ViewModelFactory.
+ * @param useCase is unique instance that manages the getting of the pokemon list data through the local or remote data sources.
+ * Because the view model receive an argument is necessary create an ViewModelFactory.
  */
-class PokemonViewModel(private val repository: PokemonRepository) : ViewModel() {
+class PokemonViewModel(private val useCase: GetAllPokemonUseCase) : ViewModel() {
 
     /**
      * We will use stateFlow for the reactive coroutines.
      * On this way we close to the recommended Android architecture guide.
-     * When we use stateFlow we must change the implementation of the observers.
+     * When we use stateFlow we must change the implementation of the observer
+     * and replace the call of observe to collect function.
      */
     private val _uiState = MutableStateFlow(UIState())
     val uiState: StateFlow<UIState> = _uiState.asStateFlow()
@@ -30,8 +32,7 @@ class PokemonViewModel(private val repository: PokemonRepository) : ViewModel() 
     init {
         viewModelScope.launch {
             _uiState.update { it.copy(loading = true) }
-            val pokeList = repository.getFirstPage()
-            _uiState.update { it.copy(pokemonList = pokeList) }
+            _uiState.update { it.copy(pokemonList = useCase()) }
             _uiState.update { it.copy(loading = false) }
         }
     }
@@ -42,8 +43,8 @@ class PokemonViewModel(private val repository: PokemonRepository) : ViewModel() 
  * In this case for the system could build the view model need the factory viewModel class.
  */
 @Suppress("UNCHECKED_CAST")
-class PokemonViewModelFactory(private val repository: PokemonRepository) : ViewModelProvider.Factory {
+class PokemonViewModelFactory(private val useCase: GetAllPokemonUseCase) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return PokemonViewModel(repository) as T
+        return PokemonViewModel(useCase) as T
     }
 }
